Added test framework
