+++ crates/deno_task_shell/Cargo.toml
@@ -29,9 +29,9 @@ pest_derive = "2.7.12"
 dirs = "5.0.1"
 pest_ascii_tree = { git = "https://github.com/prsabahrami/pest_ascii_tree.git", branch = "master" }
 miette = "7.2.0"
+tempfile = "3.12.0"
+pretty_assertions = "1.0.0"
 
 [dev-dependencies]
 parking_lot = "0.12.3"
-pretty_assertions = "1"
 serde_json = "1.0.128"
-tempfile = "3.12.0"
