+++ crates/shell/Cargo.toml
@@ -12,6 +12,10 @@ readme.workspace = true
 default-run = "shell"
 publish = false
 
+[lib]
+name = "shell"
+path = "src/lib.rs"
+
 [[bin]]
 name = "shell"
 path = "src/main.rs"
@@ -21,7 +25,7 @@ path = "src/main.rs"
 [dependencies]
 anyhow = "1.0.87"
 clap = { version = "4.5.17", features = ["derive"] }
-deno_task_shell = { path = "../deno_task_shell" }
+deno_task_shell = { path = "../deno_task_shell", features = ["shell"] }
 futures = "0.3.30"
 rustyline = { version = "14.0.0", features = ["derive"] }
 tokio = "1.40.0"
@@ -29,6 +33,8 @@ uu_ls = "0.0.27"
 dirs = "5.0.1"
 which = "6.0.3"
 uu_uname = "0.0.27"
+pretty_assertions = "1.0.0"
+tempfile = "3.12.0"
 
 [package.metadata.release]
 # Dont publish the binary
