+++ crates/shell/src/commands/mod.rs
@@ -1,4 +1,4 @@
-use std::{ffi::OsString, fs};
+use std::{collections::HashMap, ffi::OsString, fs, rc::Rc};
 
 use deno_task_shell::{EnvChange, ExecuteResult, ShellCommand, ShellCommandContext};
 use futures::{future::LocalBoxFuture, FutureExt};
@@ -20,6 +20,32 @@ pub struct AliasCommand;
 pub struct UnAliasCommand;
 
 pub struct SourceCommand;
+
+pub fn get_commands() -> HashMap<String, Rc<dyn ShellCommand>> {
+    HashMap::from([
+        ("ls".to_string(), Rc::new(LsCommand) as Rc<dyn ShellCommand>),
+        (
+            "alias".to_string(),
+            Rc::new(AliasCommand) as Rc<dyn ShellCommand>,
+        ),
+        (
+            "unalias".to_string(),
+            Rc::new(UnAliasCommand) as Rc<dyn ShellCommand>,
+        ),
+        (
+            "source".to_string(),
+            Rc::new(SourceCommand) as Rc<dyn ShellCommand>,
+        ),
+        (
+            "which".to_string(),
+            Rc::new(WhichCommand) as Rc<dyn ShellCommand>,
+        ),
+        (
+            "uname".to_string(),
+            Rc::new(UnameCommand) as Rc<dyn ShellCommand>,
+        ),
+    ])
+}
 
 impl ShellCommand for AliasCommand {
     fn execute(&self, context: ShellCommandContext) -> LocalBoxFuture<'static, ExecuteResult> {
