+++ crates/shell/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod commands;
+pub mod execute;
