+++ crates/tests/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "tests"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+deno_task_shell = { path = "../deno_task_shell", features = ["shell"] }
+shell = { path = "../shell" }
+anyhow = "1.0.87"
+futures = "0.3.30"
+pretty_assertions = "1.0.0"
+tempfile = "3.12.0"
+tokio = { version = "1.40.0", features = ["full"] }
